refactor: 로직 폴더 생성
